requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
ebaysdk==2.2.0
Pillow==10.1.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
cryptography==41.0.7 